Added budget endpoints. Cleaned up code. Added more comments. Moved helper functions to helper file.
